BUG: Use singular values instead of eigenvalues for scalar measurement
This patch aimed to address the issues explained at:
https://github.com/NIRALUser/DTIProcessToolkit/issues/4

First, FastSymmetricEigenAnalysisImageFilter is modified to return the
sorted absolute values of eigenvalues.
Also, the "OrderEigenValuesBy()" function is removed from this class,
since it was just copied from the itk class and was not used.

Second, the MD is directly computed from the eigenvalues rather than
using GetTrace() from input diffusion tensor, so there is no more
need to the "TensorMeanDiffusivityImageFilter" class.